requirements.txt
torch
torchaudio
pyannote.audio==3.1.1
scikit-learn
soundfile
tqdm
demucs
numpy==1.26.4
scipy
gradio_client
python-dotenv


